Move _NearestSiteLookup to site_collection module
The lookup type is generic (not polyhedral-specific) and will be
reused by SphericalSiteCollection. Moving it to the base module
avoids cross-imports between collection subclasses.
